+++ src/pages/transaction-signing/components/asset-item.tsx
