+++ src/pages/transaction-signing/components/error.tsx
