+++ src/pages/transaction-signing/components/popup-header.tsx
