+++ src/pages/transaction-signing/components/stx-transfer-details.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { color, Stack } from '@stacks/ui';
 import { AttachmentRow } from './attachment-row';
-import { RowItem } from '@components/transactions/row-item';
 import { Title } from '@components/typography';
 import { Divider } from '@components/divider';
-import { useTransactionRequest } from '@common/hooks/transaction/use-transaction';
+import { useTransactionRequest } from '@pages/transaction-signing/hooks/use-transaction';
+import { RowItem } from './row-item';
 
 export const StxTransferDetails: React.FC = () => {
   const pendingTransaction = useTransactionRequest();
